requirements.txt
streamlit>=1.33.0
pandas>=2.2.1
python-dotenv>=1.0.1
langchain>=0.1.0
langchain-community>=0.0.20
langchain-experimental>=0.0.47
langchain-groq>=0.0.1
langchain-huggingface>=0.0.3
langchain-core>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.6.1
tabulate==0.9.0
